+++ src/datasource/datasource.rs
@@ -9,6 +9,6 @@ use core::feature::Feature;
 
 
 pub trait DatasourceInput {
-    fn retrieve_features<F>(&self, layer: &Layer, extent: &Extent, zoom: u16, mut read: F)
+    fn retrieve_features<F>(&self, layer: &Layer, extent: &Extent, zoom: u8, mut read: F)
         where F : FnMut(&Feature);
 }
