+++ src/webserver/server.rs
@@ -182,7 +182,7 @@ pub fn webserver(args: &ArgMatches) {
         let service: &MvtService = res.server_data();
 
         let tileset = req.param("tileset").unwrap();
-        let z = req.param("z").unwrap().parse::<u16>().unwrap();
+        let z = req.param("z").unwrap().parse::<u8>().unwrap();
         let x = req.param("x").unwrap().parse::<u16>().unwrap();
         let y = req.param("y").unwrap().parse::<u16>().unwrap();
